Project/nftserver/requirements.txt
fastapi
uvicorn
pymongo
bcrypt
cryptography
fastapi-jwt-auth
python-dotenv
